# api-contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreatePlatformAccountAdjustmentRequest" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "platform_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "transaction_type" : {
            "enum" : [ "FX_IN", "FX_OUT", "FEE", "ROUNDING", "ADJUSTMENT", "MANUAL" ],
            "type" : "string"
          }
        },
        "required" : [ "amount", "platform_account_public_id", "transaction_type" ],
        "type" : "object"
      },
      "CreateTreasuryAccountRequest" : {
        "properties" : {
          "account_type" : {
            "enum" : [ "BANK_CARD", "BANK_ACCOUNT", "CRYPTO_WALLET", "EXCHANGE_ACCOUNT", "P2P_DROP", "CASH", "OTHER" ],
            "type" : "string"
          },
          "code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "currency_code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "is_active" : {
            "type" : "boolean"
          },
          "note" : {
            "type" : "string"
          },
          "title" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "account_type", "code", "currency_code", "title" ],
        "type" : "object"
      },
      "CreateTreasuryTransactionRequest" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "external_reference" : {
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "transaction_type" : {
            "enum" : [ "DEPOSIT_IN", "WITHDRAWAL_OUT", "MANUAL_IN", "MANUAL_OUT", "ADJUSTMENT", "TRANSFER_IN", "TRANSFER_OUT" ],
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "amount", "transaction_type", "treasury_account_public_id" ],
        "type" : "object"
      },
      "CreateTreasuryTransferRequest" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "external_reference" : {
            "type" : "string"
          },
          "from_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "from_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "to_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "to_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          }
        },
        "required" : [ "from_account_public_id", "from_amount", "to_account_public_id", "to_amount" ],
        "type" : "object"
      },
...
      "PagePlatformAccountTransactionResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/PlatformAccountTransactionResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PageTreasuryTransactionResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/TreasuryTransactionResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "PlatformAccountTransactionResponse" : {
        "properties" : {
          "actor_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amount" : {
            "type" : "number"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "group_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "platform_account_code" : {
            "type" : "string"
          },
          "platform_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ref_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ref_type" : {
            "type" : "string"
          },
          "transaction_type" : {
            "enum" : [ "FX_IN", "FX_OUT", "FEE", "ROUNDING", "ADJUSTMENT", "MANUAL" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "TreasuryAccountResponse" : {
        "properties" : {
          "account_type" : {
            "enum" : [ "BANK_CARD", "BANK_ACCOUNT", "CRYPTO_WALLET", "EXCHANGE_ACCOUNT", "P2P_DROP", "CASH", "OTHER" ],
            "type" : "string"
          },
          "balance" : {
            "type" : "number"
          },
          "code" : {
            "type" : "string"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "is_active" : {
            "type" : "boolean"
          },
          "note" : {
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TreasuryTransactionResponse" : {
        "properties" : {
          "actor_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amount" : {
            "type" : "number"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "external_reference" : {
            "type" : "string"
          },
          "group_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "operation_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "operation_type" : {
            "enum" : [ "DEPOSIT_REQUEST", "WITHDRAWAL_REQUEST" ],
            "type" : "string"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "transaction_type" : {
            "enum" : [ "DEPOSIT_IN", "WITHDRAWAL_OUT", "MANUAL_IN", "MANUAL_OUT", "ADJUSTMENT", "TRANSFER_IN", "TRANSFER_OUT" ],
            "type" : "string"
          },
          "treasury_account_code" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_account_title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TreasuryTransferResponse" : {
        "properties" : {
          "incoming" : {
            "$ref" : "#/components/schemas/TreasuryTransactionResponse"
          },
          "outgoing" : {
            "$ref" : "#/components/schemas/TreasuryTransactionResponse"
          }
        },
        "type" : "object"
      },
...
      "UpdateTreasuryAccountRequest" : {
        "properties" : {
          "account_type" : {
            "enum" : [ "BANK_CARD", "BANK_ACCOUNT", "CRYPTO_WALLET", "EXCHANGE_ACCOUNT", "P2P_DROP", "CASH", "OTHER" ],
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "is_active" : {
            "type" : "boolean"
          },
          "note" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Generated API contract for frontend/backend synchronization.",
    "title" : "Prime Market Backend API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/backoffice/platform-accounts/transactions" : {
      "get" : {
        "operationId" : "listTransactions_1",
        "parameters" : [ {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 30,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,DESC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagePlatformAccountTransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-platform-account-controller" ]
      },
      "post" : {
        "operationId" : "createAdjustment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePlatformAccountAdjustmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlatformAccountTransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-platform-account-controller" ]
      }
    },
    "/api/backoffice/treasury/accounts" : {
      "get" : {
        "operationId" : "listAccounts",
        "parameters" : [ {
          "in" : "query",
          "name" : "active_only",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TreasuryAccountResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-treasury-controller" ]
      },
      "post" : {
        "operationId" : "createAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTreasuryAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TreasuryAccountResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-treasury-controller" ]
      }
    },
    "/api/backoffice/treasury/accounts/{publicId}" : {
      "patch" : {
        "operationId" : "updateAccount",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTreasuryAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TreasuryAccountResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-treasury-controller" ]
      }
    },
...
    "/api/backoffice/treasury/transactions" : {
      "get" : {
        "operationId" : "listTransactions",
        "parameters" : [ {
          "in" : "query",
          "name" : "account_public_id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 30,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,DESC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageTreasuryTransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-treasury-controller" ]
      },
      "post" : {
        "operationId" : "createTransaction",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTreasuryTransactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TreasuryTransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-treasury-controller" ]
      }
    },
    "/api/backoffice/treasury/transfers" : {
      "post" : {
        "operationId" : "createTransfer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTreasuryTransferRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TreasuryTransferResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-treasury-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Relative to the current backend base URL",
    "url" : "/"
  } ]
}